+++ test/grid-size-tests.js
@@ -363,3 +363,4 @@ function runAllTests() {
     return passed === total;
 }
 
+